agents/sydney/requirements.txt
livekit-agents[openai,deepgram,cartesia,silero,turn-detector]~=1.5
livekit-plugins-noise-cancellation~=0.2
livekit-api~=1.0
python-dotenv~=1.0
pytest~=8.0
pytest-asyncio~=0.24
